FS-2: Split controller
